refactor(minimald): move minimald RPC types to own crate
