+++ crates/minimald-rpc/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "minimald-rpc"
+version = "0.0.1"
+edition.workspace = true
+publish.workspace = true
+
+[dependencies]
+constcat.workspace = true
+serde.workspace = true
+
+sessions.workspace = true
